backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyarrow>=14.0.0
